cargo: Build deps with release profile, use mold and compressed debug symbols
Build dependencies with optimizations. Given that dependency compilation happens very infrequently this gives us most of the benefit of optimizations without increasing build times in the most common case of a developer rebuilding the project after making code changes (but not dependency changes).

Also bring back compressed debug symbols as an alternative for shrinking the binary with "line-tables-only". This shrinks the binary further than line-tables-only but without interfering with the ability to attach a debugger. To get back some of the lost linker time switch to using the mold linker.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>