moss: Add origin field to package meta
This is admittedly a bit of a hack in order to get an origin from
a package, without reorchestrating how repos work. Importantly, it is
not added to the moss db.

The origin of packages is fundamental to how package ids work in
packagekit which consist of name:version:arch:data.

the data field may look like:
"installed:volatile" for installed packages with a origin
"volatile" if available and not installed
"installed" if installed without a repo origin

These package ids get passed everywhere in packagekit so it is important
they are constructed properly.