boulder: Adjust patch again
The previous flags didn't _quite_ do what I wanted, so switch `--batch` back to `--force` so that patch won't attempt to reverse patches that have already been flagged. The `--forward` flag does the same thing and is technically redundant with `--force` but it makes it clear what patch should be doing.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>