boulder: Eliminate all lingering threads
After spending some time in `lldb` to figure out
what threads still exist when we enter the container,
I found the following:

- zbus "blocking" wraps its async client and uses
the `async-io` crate to `block_on` that client. This
created a single global runtime thread that we had
no control over. I've changed this to opt-in to its
`tokio` feature so it won't spawn its own runtime,
and modified `runtime::block_on` to allow things
to get called when no global (`runtime::init`) runtime
is initialized by running it on a temporary runtime.
In either case, all threads are now dropped.

- rayon spawns its own global runtime we have no control
over. I've updated the `par_iter` invocation in the
boulder call path to use a temporary runtime that we
can drop

All threads are now dropped before we CLONE into the
user namespace ("container") which should hopefully
eliminate the risk of us sometimes deadlocking