moss: Add system model MVP
This adds an initial MVP of system model

## Imperative

We define "imperative" usage of moss as the existing
workflow. Users issue consequtive commands to mutate
the state of their system. During `sync`, the active
state is the source of truth. If you were to `remove`
a package, that new state is the source of truth.

## Declarative (`system-model`)

System model is a _declarative_ way to define the
desired state of a system. A system model file
explicitly defines the repositories & packages a
user wants to to `sync` their system against. During
`sync`, the system model file is the source of truth
instead of the active state, like in "imperative" mode.

`install`, `remove`, `state active`, etc can still
be used to imperatively mutate the active state, but
unless the desired changes are written to the system
model file, these changes are temporary and will be
lost on the next `sync`.

Users _opt in_ to using system model by writing a
`{root}/etc/moss/system-model.kdl` file.

## system-model.kdl

A very simple implementation of defining repositories
and packages via KDL is implemented and looks like:

```kdl
repositories {
    volatile {
        uri "https://infratest.aerynos.dev/vessel/volatile/x86_64/stone.index"
        priority 1
    }
}
packages {
    bash
    nano {
      why "vim scares me"
    }
    "pkgconfig(zlib)"
    "binary(cargo)"
}
```

## TODO

There a lots of things still left to do, like
flesh out the full system-model.kdl format and
ensure `moss` UX is consistent and accurate across
all commands when `sysmtem-model` is active, such
as disabling commands to add / remove repos, since
that needs to be done explicitly in the system model
file.