boulder: Add global cargo cache
The cargo global cache, used for caching downloaded crates and dependencies, goes in $CARGO_HOME, so set that value and add a persistent mapping for it.

This was tested to bring the fetch phase for amdgpu_top down from 5.09s to 0.63s.

While that gain is marginal for single-package builds the real value is when doing a large rust stack like COSMIC where many dependencies are shared (though the impact is still likely minimal).

Also, while I didn't test any other packages in the repo I can forsee that some packages have unique CARGO_HOME handling and may break some builds. To indicate where that happens add a descriptive macro that unsets the variable.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>